+++ x/fungible/keeper/evm_gateway_test.go
@@ -28,7 +28,7 @@ func TestKeeper_DepositAndCallZeta(t *testing.T) {
 		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
 		_ = deployZRC20(t, ctx, k, sdkk.EvmKeeper, chainID, "foo", sample.EthAddress().String(), "foo")
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)
 
@@ -58,7 +58,7 @@ func TestKeeper_DepositAndCallZeta(t *testing.T) {
 		message := []byte("test message")
 		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)
 
@@ -92,7 +92,7 @@ func TestKeeper_DepositAndCallZeta(t *testing.T) {
 			SystemContract: sample.EthAddress().Hex(),
 		})
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)
 
@@ -127,7 +127,7 @@ func TestKeeper_CallDepositAndRevert(t *testing.T) {
 		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
 		zrc20 := deployZRC20(t, ctx, k, sdkk.EvmKeeper, chainID, "foo", sample.EthAddress().String(), "foo")
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)
 
@@ -175,7 +175,7 @@ func TestKeeper_CallDepositAndRevert(t *testing.T) {
 		message := []byte("test message")
 		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)
 
@@ -206,7 +206,7 @@ func TestKeeper_CallDepositAndRevert(t *testing.T) {
 			SystemContract: sample.EthAddress().Hex(),
 		})
 
-		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM)
+		testDAppV2, err := k.DeployContract(ctx, testdappv2.TestDAppV2MetaData, true, types.ModuleAddressEVM, sample.EthAddress())
 		require.NoError(t, err)
 		require.NotEmpty(t, testDAppV2)