+++ changelog.md
@@ -2,9 +2,17 @@
 
 ## Unreleased
 
+### Features
+
+* [4064](https://github.com/zeta-chain/node/pull/4064) - add support for withdraws using the new v2 connector contract
+
 ### Refactor
 
 * [4070](https://github.com/zeta-chain/node/pull/4070) - remove support for v1 revert address for BTC
+
+### Fixes
+
+* [4090](https://github.com/zeta-chain/node/pull/4090) - print error message in detail if unable to decode Bitcoin memo
 
 ## v33.0.0